hmpsentinel_v3/requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pdfminer.six>=20231228
python-dateutil>=2.8.2
requests>=2.31.0
